Move valibot error formatter into util directory
